Split ivpsolve into subpackages
